requirements.txt
Flask==2.2.5
gunicorn==20.1.0
tensorflow-cpu==2.20.0
opencv-python-headless==4.8.0.76
numpy==1.26.4
pillow==10.0.0
pandas==2.2.2
scikit-learn==1.3.0
